一旦コミット

+++ src/main/java/me/hannsi/lfjg/render/openGL/system/Id.java
@@ -2,6 +2,6 @@ package me.hannsi.lfjg.render.openGL.system;
 
 public class Id {
     public static final long initialGLObjectId = Long.MIN_VALUE;
-    public static final long initialEffectCacheId = Integer.MIN_VALUE;
+    public static final long initialEffectCacheId = Long.MIN_VALUE;
     public static long glLatestObjectId = initialGLObjectId;
 }
